refactor: 마이페이지 구조 정리 및 단일 라우트 통합
