--- src/app/myPage/teacher/page.tsx
@@ -1,6 +1,0 @@
-import MyPageScreen from '@/components/myPage/MyPageScreen';
-import { teacherProfile } from '@/components/myPage/mockData';
-
-export default function TeacherMyPage() {
-  return <MyPageScreen profile={teacherProfile} currentRole="teacher" />;
-}
