--- src/app/myPage/user/page.tsx
@@ -1,6 +1,0 @@
-import MyPageScreen from '@/components/myPage/MyPageScreen';
-import { userProfile } from '@/components/myPage/mockData';
-
-export default function UserMyPage() {
-  return <MyPageScreen profile={userProfile} currentRole="user" />;
-}
