+++ src/components/myPage/ExhibitionList.tsx
