+++ src/components/myPage/NewExhibitionBanner.tsx
@@ -1,4 +1,4 @@
-import Image from 'next/image';
+import { Palette } from 'lucide-react';
 import Link from 'next/link';
 
 export default function NewExhibitionBanner() {
@@ -6,13 +6,7 @@ export default function NewExhibitionBanner() {
     <section className="flex items-center justify-between rounded-[22px] bg-[linear-gradient(90deg,#f0b63b_0%,#ff8d73_100%)] px-5 py-5 shadow-[0_8px_18px_rgba(247,154,100,0.18)]">
       <div>
         <div className="mb-1.5 flex items-center gap-2">
-          <Image
-            src="/Icon.png"
-            alt="새 전시회 아이콘"
-            width={18}
-            height={18}
-            className="shrink-0"
-          />
+          <Palette size={18} className="shrink-0 text-white" />
           <span className="text-[15px] font-bold text-white">
             새 전시회 만들기
           </span>
